Add investment planner pages
