+++ README.md
@@ -1,0 +1,12 @@
+# Kids Investment Site
+
+A simple Next.js style project with a few pages to help parents plan investments for their children.
+
+## Pages
+
+- **Home** (`/`): Links to tools.
+- **Savings Calculator** (`/calculator`): Estimate future savings based on monthly contributions.
+- **529 Plan Advantages** (`/529`): See state-specific tax benefits for 529 plans.
+- **401k Info** (`/401k`): Learn when a child can start a 401k.
+
+These pages use basic React components and minimal sample data. You can run them in any Next.js project by placing the `pages` folder in your app.
